Moving fixtures into brewtils.test
